frozen_requirements.txt
alabaster==0.7.16
amqp==5.2.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
babel==2.15.0
billiard==4.2.0
cachecontrol==0.14.0
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.0
crispy-bootstrap5==2024.2
cron-descriptor==1.4.3
cryptography==43.0.0
cssbeautifier==1.15.1
decorator==5.1.1
distlib==0.3.8
dj-rest-auth==6.0.0
django==5.0.8
django-allauth==64.0.0
django-anymail==11.0.1
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-coverage-plugin==3.1.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-jazzmin==3.0.0
django-model-utils==4.5.1
django-multiselectfield==0.1.13
django-redis==5.4.0
django-stubs==5.0.4
django-stubs-ext==5.0.4
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.0
djlint==1.34.1
docutils==0.21.2
drf-spectacular==0.27.2
drf-standardized-errors==0.14.0
editorconfig==0.12.4
exceptiongroup==1.2.2
executing==2.0.1
factory-boy==3.3.0
faker==26.1.0
fido2==1.1.3
filelock==3.15.4
firebase-admin==6.5.0
flower==2.0.1
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
hiredis==3.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httplib2==0.22.0
httptools==0.6.1
humanize==4.10.0
identify==2.6.0
idna==3.7
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.26.0
jedi==0.19.1
jinja2==3.1.4
jsbeautifier==1.15.1
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
markupsafe==2.1.5
matplotlib-inline==0.1.7
msgpack==1.0.8
mypy==1.11.1
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
parso==0.8.4
pathspec==0.12.1
paypalrestsdk==1.13.3
pexpect==4.9.0
pillow==10.4.0
pip==24.2
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
prometheus-client==0.20.0
prompt-toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psycopg==3.2.1
psycopg-c==3.2.1
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pygments==2.18.0
pyjwt==2.9.0
pyopenssl==24.2.1
pyparsing==3.1.2
pypng==0.20220715.0
pytest==8.3.2
pytest-django==4.8.0
pytest-sugar==1.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
pyyaml==6.0.1
qrcode==7.4.2
redis==5.0.8
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
rest-framework-simplejwt==0.0.2
rpds-py==0.19.1
rsa==4.9
ruff==0.5.6
ruff-pre-commit @ file:///home/amitpatel/.cache/pre-commit/repov17sevy5
setuptools==70.2.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sphinx==7.4.7
sphinx-autobuild==2024.4.16
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.38.2
termcolor==2.4.0
text-unidecode==1.3
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
types-pyyaml==6.0.12.20240724
types-requests==2.32.0.20240712
typing-extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uv==0.4.14
uvicorn==0.30.4
uvicorn-worker==0.2.0
uvloop==0.19.0
vine==5.1.0
virtualenv==20.26.3
watchdog==4.0.1
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
werkzeug==3.0.3
wheel==0.43.0
whitenoise==6.7.0
